fixed collection tests, organized imports
