+++ ancile/core/decorators/advanced/aggregate_decorator.py
@@ -1,9 +1,7 @@
-import wrapt
 from ancile.core.decorators import *
 import logging
 
-from ancile.core.primitives import DataPolicyPair, Collection, Policy
-from ancile.core.primitives.policy_helpers.expressions import IntersectExpression
+from ancile.core.primitives import *
 from ancile.core.user_secrets import UserSecrets
 from ancile.utils.errors import AncileException
