+++ ancile/core/decorators/advanced/collection/filter_decorator.py
@@ -1,6 +1,7 @@
 import wrapt
 from ancile.core.decorators import *
 import logging
+
 logger = logging.getLogger(__name__)
 
 
@@ -10,8 +11,21 @@ class FilterDecorator(BaseDecorator):
         super().__init__(scopes, is_collection)
         self.scopes.append('filter')
 
-    @staticmethod
-    def process_call(command, is_collection):
-        logger.debug('Calling Filter Decorator')
+    def process_call(self, command):
 
         return NotImplemented
+        # I believe it's already in Collection class
+
+        # logger.debug('Calling Filter Decorator')
+        # f = command.params.get("lambda_function", False)
+        # collection = command.params.get("collection", False)
+        # if not f and collection:
+        #     raise AncileException("For filter function")
+        # new_collection = Collection()
+        # for data_point in collection._data_points:
+        #     value = f(data_point)
+        #     if value:
+        #         data_point._advance_policy('filter_keep')
+        #         new_collection._data_points.append(data_point)
+        #     else:
+        #         data_point._advance_policy('filter_remove')
