+++ ancile/core/decorators/advanced/comparison_decorator.py
@@ -1,9 +1,6 @@
-import wrapt
 import logging
-import copy
 from ancile.core.decorators import *
-from ancile.core.primitives import DataPolicyPair
-from ancile.core.primitives.policy_helpers.command import Command
+from ancile.core.primitives import *
 
 logger = logging.getLogger(__name__)
