+++ ancile/core/decorators/external_decorator.py
@@ -1,6 +1,4 @@
-import wrapt
 import logging
-import copy
 import inspect
 from ancile.core.decorators import *
 from ancile.core.user_secrets import UserSecrets
