+++ ancile/core/primitives/__init__.py
@@ -1,5 +1,5 @@
+from ancile.core.primitives.data import Data
+from ancile.core.primitives.command import Command
+from ancile.core.primitives.policy import Policy
 from ancile.core.primitives.data_policy_pair import DataPolicyPair
 from ancile.core.primitives.collection import Collection
-from ancile.core.primitives.policy import Policy
-from ancile.core.primitives.result import Result
-from ancile.core.primitives.data import Data
