+++ ancile/core/primitives/policy.py
@@ -1,6 +1,6 @@
 from __future__ import annotations
 from copy import deepcopy
-from ancile.core.primitives.policy_helpers.command import Command
+from ancile.core.primitives.command import Command
 from ancile.core.primitives.policy_helpers.policy_parser import PolicyParser
 from ancile.core.primitives.policy_helpers.expressions import *
